roles and permissions setup and seeder

+++ app/Http/Controllers/UserController.php
@@ -10,6 +10,7 @@ use App\Models\User;
 use DataTables;
 use Illuminate\Validation\Rule;
 use Illuminate\Validation\Rules;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
@@ -20,11 +21,12 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('auth.register');
+        $roles = Role::all();
+        return view('auth.register', compact('roles'));
     }
 
     /**
-     * Handle an incoming registration request.
+     * Handle an incoming registration request with assigned role.
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\RedirectResponse
@@ -33,6 +35,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        // dd($request);
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
@@ -44,26 +47,49 @@ class UserController extends Controller
             'email' => $request->email,
             'password' => Hash::make($request->password),
         ]);
+        $user->assignRole($request->role);
 
         return redirect('/data');
     }
 
+    /**
+     * Display the edit view with associated user 
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\RedirectResponse
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     */
     public function edit(Request $request)
     {
+        $roles = Role::all();
         $user = User::findOrFail($request->id);
-        return view('edit', compact('user'));
+        return view('edit', compact(['user', 'roles']));
     }
 
+    /**
+     * Handle an incoming update request with assigned role.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\RedirectResponse
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     */
     public function update(Request $request)
     {
         $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => Rule::unique('users')->ignore($request->id),
         ]);
-        $user = User::findOrFail($request->id);  
-        $user->update($request->all());
+        $user = User::findOrFail($request->id);
+        $user->update([
+            'name' => $request->name,
+            'email' => $request->email,
+        ]);
+        $role = Role::where('id', $request->role)->first();
+        $user->syncRoles($role);
 
-        return redirect('/data');
+        return redirect('/users');
     }
 
     public function destroy($id)
@@ -81,9 +107,16 @@ class UserController extends Controller
             return Datatables::of($users)
                 ->addIndexColumn()
                 ->addColumn('action', function($user){
-                    $actionBtn = '<a href="http://localhost:8000/edit/user/'.$user->id .'" class="edit btn btn-success btn-sm">Edit</a> 
-                    <a href="http://localhost:8000/delete/user/'.$user->id .'" class="delete btn btn-danger btn-sm">Delete</a>';
-                    return $actionBtn;
+                    $auth_user = Auth::user();
+                    if($auth_user->can('edit users') || $auth_user->can('delete users')){
+                        $actionBtn = '<a href="http://localhost:8000/edit/user/'.$user->id .'" class="edit btn btn-success btn-sm">Edit</a> 
+                        <a href="http://localhost:8000/delete/user/'.$user->id .'" class="delete btn btn-danger btn-sm">Delete</a>';
+                        return $actionBtn;
+                    }
+                    
+                    $actionBtn = '<a href="#" class="edit btn disabled btn-success btn-sm">Edit</a> 
+                        <a href="#" class="delete btn disabled btn-danger btn-sm">Delete</a>';
+                        return $actionBtn;
                 })
                 ->rawColumns(['action'])
                 ->make(true);
